requirements.txt
numpy
Pillow
scipy
scikit-image
scikit-learn
tqdm
